refactor: PR issue fix
