+++ src/components/ui/deck-list-table/deck-list-table.module.scss
